* commit 'c829185dc0b5ea77c1cfae03743a7c476afdd678':
  provider: avoid potential unaligned reads/writes
  treewide: abandon static_cast<long> and use %z
  treewide: add const to operator() functions
  treewide: replace sizeof(char) by 1
  freebusy: move non-KC interfaces out of KC namespace
  freebusy: move private struct out of public header
  client: remove unused IID_ISelectUnicode
  client: remove unused IIDs for {AB,MS}Provider

+++ provider/client/ECExchangeExportChanges.cpp
@@ -958,8 +958,8 @@ HRESULT ECExchangeExportChanges::ExportMessageChangesFast()
 	const auto lpImportProps = m_sourcekey.empty() ? sptImportPropsServerWide : sptImportProps;
 
 	// No more changes (add/modify).
-	ZLOG_DEBUG(m_lpLogger, "ExportFast: At step %u, changeset contains %lu items)",
-		m_ulStep, static_cast<unsigned long>(m_lstChange.size()));
+	ZLOG_DEBUG(m_lpLogger, "ExportFast: At step %u, changeset contains %zu items)",
+		m_ulStep, m_lstChange.size());
 	if (m_ulStep >= m_lstChange.size())
 		goto exit;
